backend/mutual_fund_broker/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
Django==4.2.18
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-rest-swagger==2.2.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
kombu==5.4.2
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
prompt_toolkit==3.0.48
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
